requirements-py3.txt
altgraph==0.15
antlr4-python3-runtime==4.7.1
attrs==17.4.0
bidict==0.15.0
flatty==0.1.2
future==0.16.0
macholib==1.9
more-itertools==4.1.0
pefile==2017.11.5
pluggy==0.6.0
py==1.5.3
PyInstaller==3.3.1
pytest==3.5.0
six==1.11.0
StringBuilder==1.0.0a4
untangle==1.1.1
